templates

+++ Templates/My_Templates/Graphs/Djisktra.dart
@@ -3,10 +3,10 @@ vector<int>parent; // If we need to retrace path just backtrack
 // Time Complexity : O(E + V logV)
 
 // For distance between a given node and any other node, edges>=0
-vector <ll> dijkstra(vector<vector<pair<int,ll>>>&adj, vector<int>&src){
+vector <int> dijkstra(vector<vector<pair<int,int>>>&adj, vector<int>&src){
     int n = adj.size();
-    vector<ll>dist(n,inf);
-    priority_queue<pair<ll,int>>pq;
+    vector<int>dist(n,inf);
+    priority_queue<pair<int,int>>pq;
     for(auto source : src){
         dist[source] = 0;
         pq.push({0,source});
@@ -14,14 +14,14 @@ vector <ll> dijkstra(vector<vector<pair<int,ll>>>&adj, vector<int>&src){
     
     while(!pq.empty()){
         auto x = pq.top();
-        ll d = -1*x.first; int node = x.second;
+        int d = -1*x.first; int node = x.second;
         pq.pop();
         
         if(d!=dist[node]){ // VERY IMPORTANT (PREVENTS TLE)
             continue;
         }
         for(auto child : adj[node]){
-            int v = child.first; ll w = child.second;
+            int v = child.first; int w = child.second;
             if((d+w)<dist[v]){
                 // parent[v]  = node;
                 dist[v] = d+w;
