+++ Templates/My_Templates/Graphs/UFDS+DSU.dart
@@ -1,18 +1,18 @@
 struct ufds{
-    vector<ll>par; 
-    vector<ll>sz; 
-    ufds(ll n){
+    vector<int>par; 
+    vector<int>sz; 
+    ufds(int n){
         par.resize(n); 
         sz.resize(n);
-        for(ll i = 0;i<n;i++){
+        for(int i = 0;i<n;i++){
             par[i] = i; sz[i] = 1;
         }
     }
-    ll get(ll u){
+    int get(int u){
         if(u==par[u]) return u;
         return par[u] = get(par[u]);
     }
-    void unite(ll u, ll v){
+    void unite(int u, int v){
         //u--; v--;
         u = get(u); v = get(v);
         if(u!=v){
